+++ tests/test_exceptions.py
@@ -1,25 +1,25 @@
 import unittest
 from ._compat import patch
 
-from telegram_upload.exceptions import TelegramUploadError, catch
+from telegram_upload.exceptions import dxdmgchUploadError, catch
 
 
-class TestTelegramUploadError(unittest.TestCase):
+class TestdxdmgchUploadError(unittest.TestCase):
     def test_exception(self):
-        self.assertEqual(str(TelegramUploadError()), 'TelegramUploadError')
+        self.assertEqual(str(dxdmgchUploadError()), 'dxdmgchUploadError')
 
     def test_body(self):
-        error = TelegramUploadError()
+        error = dxdmgchUploadError()
         error.body = 'body'
-        self.assertEqual(str(error), 'TelegramUploadError: body')
+        self.assertEqual(str(error), 'dxdmgchUploadError: body')
 
     def test_extra_body(self):
-        self.assertEqual(str(TelegramUploadError('extra_body')), 'TelegramUploadError: extra_body')
+        self.assertEqual(str(dxdmgchUploadError('extra_body')), 'dxdmgchUploadError: extra_body')
 
     def test_all(self):
-        error = TelegramUploadError('extra_body')
+        error = dxdmgchUploadError('extra_body')
         error.body = 'body'
-        self.assertEqual(str(error), 'TelegramUploadError: body. extra_body')
+        self.assertEqual(str(error), 'dxdmgchUploadError: body. extra_body')
 
 
 class TestCatch(unittest.TestCase):
@@ -29,7 +29,7 @@ class TestCatch(unittest.TestCase):
     @patch('telegram_upload.exceptions.sys.stderr.write')
     def test_raise(self, m):
         def raise_error():
-            raise TelegramUploadError('Error')
+            raise dxdmgchUploadError('Error')
         with self.assertRaises(SystemExit):
             catch(raise_error)()
         m.assert_called_once()